+++ src/Runner/Runner/WithShrinking.php
@@ -56,6 +56,7 @@ final class WithShrinking
         Assert $assert,
         Value $scenario,
     ): void {
+        $identity = $this->hash($previousFailure);
         $previousStrategy = $scenario;
         $dichotomy = $scenario->shrink();
 
@@ -67,9 +68,36 @@ final class WithShrinking
             $currentStrategy = $dichotomy->a();
 
             try {
-                $currentStrategy->unwrap()($assert);
+                try {
+                    $currentStrategy->unwrap()($assert);
+                } catch (Assert\Failure $e) {
+                    if ($this->canShrink($e, $identity)) {
+                        // throwing here will go call the shrink method in the
+                        // catch below
+                        throw $e;
+                    }
+
+                    // if it can't shrink due to error identity change then we
+                    // still try the B strategy to have a more fine grained
+                    // shrunk value
+                }
                 $currentStrategy = $dichotomy->b();
-                $currentStrategy->unwrap()($assert);
+
+                try {
+                    $currentStrategy->unwrap()($assert);
+                } catch (Assert\Failure $e) {
+                    if ($this->canShrink($e, $identity)) {
+                        // throwing here will go call the shrink method in the
+                        // catch below
+                        throw $e;
+                    }
+
+                    // if the error identity changes with the B strategy then it
+                    // means we reached the most fine grained shrunk value that
+                    // still correspond to the initial error.
+                    // Here we do nothing meaning we fall through the case below
+                    // where we assign $dichotomy to null to stop the shrinking.
+                }
 
                 // When a and b work then we assign this variable to null so at
                 // the next loop iteration it'll throw the exception above
@@ -85,5 +113,24 @@ final class WithShrinking
             }
             // we can use an infinite condition here since all exits are covered
         } while (true);
+    }
+
+    private function canShrink(Assert\Failure $e, string $identity): bool
+    {
+        return $this->hash($e) === $identity;
+    }
+
+    private function hash(Assert\Failure $e): string
+    {
+        $trace = $e->getTrace();
+        /** @var array{file: string, line: string} */
+        $kind = $trace[0];
+        /** @var array{file: string, line: string} */
+        $userCall = $trace[1];
+
+        return \hash(
+            'xxh3',
+            $kind['file'].$kind['line'].$userCall['file'].$userCall['line'],
+        );
     }
 }